Get rid of the p2 folder
This commit eliminates one of traces of the prototyping process: the p2 folder.
