+++ f2/src/ContactInfoPage.js
@@ -2,13 +2,13 @@
 import { jsx } from '@emotion/core'
 import { Route } from 'react-router-dom'
 import { Trans } from '@lingui/macro'
-import { H1 } from '../components/header'
-import { P } from '../components/paragraph'
-import { Steps } from '../components/stepper'
-import { TrackPageViews } from '../TrackPageViews'
+import { H1 } from './components/header'
+import { P } from './components/paragraph'
+import { Steps } from './components/stepper'
+import { TrackPageViews } from './TrackPageViews'
 import { ContactInfoForm } from './forms/ContactInfoForm'
-import { Layout } from '../components/layout'
-import { BackButton } from '../components/backbutton'
+import { Layout } from './components/layout'
+import { BackButton } from './components/backbutton'
 
 export const ContactInfoPage = () => (
   <Route
