+++ f2/src/LandingPage.js
@@ -3,14 +3,14 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { ApolloConsumer } from 'react-apollo'
 import { Trans } from '@lingui/macro'
-import { P } from '../components/paragraph'
-import { H1 } from '../components/header'
-import { Ul } from '../components/unordered-list'
-import { Li } from '../components/list-item'
-import { InfoCard } from '../components/container'
-import { Layout } from '../components/layout'
-import { TrackPageViews } from '../TrackPageViews'
-import { ButtonsContainer } from '../components/buttons-container'
+import { P } from './components/paragraph'
+import { H1 } from './components/header'
+import { Ul } from './components/unordered-list'
+import { Li } from './components/list-item'
+import { InfoCard } from './components/container'
+import { Layout } from './components/layout'
+import { TrackPageViews } from './TrackPageViews'
+import { ButtonsContainer } from './components/buttons-container'
 
 export const LandingPage = props => {
   return (
