+++ f2/src/NextSteps.js
@@ -2,14 +2,14 @@
 import { jsx } from '@emotion/core'
 import { useLingui } from '@lingui/react'
 import { Trans } from '@lingui/macro'
-import { H1 } from '../components/header'
-import { P } from '../components/paragraph'
-import { A } from '../components/link'
-import { Ul } from '../components/unordered-list'
-import { Li } from '../components/list-item'
-import { InfoCard } from '../components/container'
-import { TrackPageViews } from '../TrackPageViews'
-import { Layout } from '../components/layout'
+import { H1 } from './components/header'
+import { P } from './components/paragraph'
+import { A } from './components/link'
+import { Ul } from './components/unordered-list'
+import { Li } from './components/list-item'
+import { InfoCard } from './components/container'
+import { TrackPageViews } from './TrackPageViews'
+import { Layout } from './components/layout'
 import { ApolloConsumer } from 'react-apollo'
 
 export const NextSteps = () => {
