+++ f2/src/ScammerDetailsPage.js
@@ -2,14 +2,14 @@
 import { jsx } from '@emotion/core'
 import { Route } from 'react-router-dom'
 import { Trans } from '@lingui/macro'
-import { H1 } from '../components/header'
-import { P } from '../components/paragraph'
-import { Steps } from '../components/stepper'
-import { TrackPageViews } from '../TrackPageViews'
+import { H1 } from './components/header'
+import { P } from './components/paragraph'
+import { Steps } from './components/stepper'
+import { TrackPageViews } from './TrackPageViews'
 import { ScammerDetailsForm } from './forms/ScammerDetailsForm'
-import { Layout } from '../components/layout'
-import { getDoneForms } from '../utils/queriesAndMutations'
-import { BackButton } from '../components/backbutton'
+import { Layout } from './components/layout'
+import { getDoneForms } from './utils/queriesAndMutations'
+import { BackButton } from './components/backbutton'
 
 export const ScammerDetailsPage = () => (
   <Route
