+++ f2/src/WhatHappenedPage.js
@@ -2,16 +2,16 @@
 import { jsx } from '@emotion/core'
 import { Route } from 'react-router-dom'
 import { Trans } from '@lingui/macro'
-import { H1 } from '../components/header'
-import { P } from '../components/paragraph'
-import { Ul } from '../components/unordered-list'
-import { Li } from '../components/list-item'
-import { Steps } from '../components/stepper'
-import { TrackPageViews } from '../TrackPageViews'
+import { H1 } from './components/header'
+import { P } from './components/paragraph'
+import { Ul } from './components/unordered-list'
+import { Li } from './components/list-item'
+import { Steps } from './components/stepper'
+import { TrackPageViews } from './TrackPageViews'
 import { WhatHappenedForm } from './forms/WhatHappenedForm'
-import { Layout } from '../components/layout'
-import { getDoneForms } from '../utils/queriesAndMutations'
-import { BackButton } from '../components/backbutton'
+import { Layout } from './components/layout'
+import { getDoneForms } from './utils/queriesAndMutations'
+import { BackButton } from './components/backbutton'
 
 export const WhatHappenedPage = () => (
   <Route
