+++ f2/src/__tests__/ConfirmationSummary.test.js
@@ -1,13 +1,13 @@
 import React from 'react'
-import { i18n } from '@lingui/core'
 import { render, cleanup } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { TimeFramePage } from '../TimeFramePage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { i18n } from '@lingui/core'
+import { ConfirmationSummary } from '../ConfirmationSummary'
+import theme from '../theme'
+import en from '../locales/en.js'
 
 i18n.load('en', { en })
 i18n.activate('en')
@@ -15,11 +15,15 @@ i18n.activate('en')
 const client = {
   readQuery: () => ({
     timeFrame: JSON.stringify({}),
+    whatHappened: JSON.stringify({}),
+    scammerDetails: JSON.stringify({}),
+    impact: JSON.stringify({}),
+    contactInfo: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<TimeFramePage />', () => {
+describe('<ConfirmationSummary />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -28,7 +32,7 @@ describe('<TimeFramePage />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <TimeFramePage />
+              <ConfirmationSummary />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
