+++ f2/src/__tests__/ContactInfoPage.test.js
@@ -6,8 +6,8 @@ import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
 import { ContactInfoPage } from '../ContactInfoPage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import theme from '../theme'
+import en from '../locales/en.js'
 
 i18n.load('en', { en })
 i18n.activate('en')
