+++ f2/src/__tests__/ImpactStatementPage.test.js
@@ -1,24 +1,25 @@
 import React from 'react'
-import { render, cleanup } from '@testing-library/react'
-import { setupI18n } from '@lingui/core'
+import { i18n } from '@lingui/core'
 import { MemoryRouter } from 'react-router-dom'
+import { render, cleanup } from '@testing-library/react'
 import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { ScammerDetailsPage } from '../ScammerDetailsPage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { ImpactStatementPage } from '../ImpactStatementPage'
+import theme from '../theme'
+import en from '../locales/en.js'
 
-const i18n = setupI18n({ catalogs: { en } })
+i18n.load('en', { en })
+i18n.activate('en')
 
 const client = {
   readQuery: () => ({
-    scammerDetails: JSON.stringify({}),
+    impact: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<ScammerDetailsPage />', () => {
+describe('<ImpactStatementPage />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -27,7 +28,7 @@ describe('<ScammerDetailsPage />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <ScammerDetailsPage />
+              <ImpactStatementPage />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
