+++ f2/src/__tests__/LandingPage.test.js
@@ -1,25 +1,29 @@
 import React from 'react'
 import { i18n } from '@lingui/core'
 import { render, cleanup } from '@testing-library/react'
-import { MemoryRouter } from 'react-router-dom'
 import { ApolloProvider } from 'react-apollo'
+import { MemoryRouter } from 'react-router-dom'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { WhatHappenedPage } from '../WhatHappenedPage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { LandingPage } from '../LandingPage'
+import theme from '../theme'
+import en from '../locales/en.js'
 
 i18n.load('en', { en })
 i18n.activate('en')
 
 const client = {
   readQuery: () => ({
-    whatHappened: JSON.stringify({}),
+    scamInfo: JSON.stringify({}),
+    lostMoney: JSON.stringify({}),
+    suspectInfo: JSON.stringify({}),
+    files: [],
+    contactInfo: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<WhatHappenedPage />', () => {
+describe('<LandingPage />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -28,7 +32,7 @@ describe('<WhatHappenedPage />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <WhatHappenedPage />
+              <LandingPage location={{ search: undefined }} />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
