+++ f2/src/__tests__/ScammerDetailsPage.test.js
@@ -1,29 +1,24 @@
 import React from 'react'
-import { i18n } from '@lingui/core'
 import { render, cleanup } from '@testing-library/react'
-import { ApolloProvider } from 'react-apollo'
+import { setupI18n } from '@lingui/core'
 import { MemoryRouter } from 'react-router-dom'
+import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { LandingPage } from '../LandingPage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { ScammerDetailsPage } from '../ScammerDetailsPage'
+import theme from '../theme'
+import en from '../locales/en.js'
 
-i18n.load('en', { en })
-i18n.activate('en')
+const i18n = setupI18n({ catalogs: { en } })
 
 const client = {
   readQuery: () => ({
-    scamInfo: JSON.stringify({}),
-    lostMoney: JSON.stringify({}),
-    suspectInfo: JSON.stringify({}),
-    files: [],
-    contactInfo: JSON.stringify({}),
+    scammerDetails: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<LandingPage />', () => {
+describe('<ScammerDetailsPage />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -32,7 +27,7 @@ describe('<LandingPage />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <LandingPage location={{ search: undefined }} />
+              <ScammerDetailsPage />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
