+++ f2/src/__tests__/TimeFramePage.test.js
@@ -1,25 +1,25 @@
 import React from 'react'
 import { i18n } from '@lingui/core'
-import { MemoryRouter } from 'react-router-dom'
 import { render, cleanup } from '@testing-library/react'
+import { MemoryRouter } from 'react-router-dom'
 import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { ImpactStatementPage } from '../ImpactStatementPage'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { TimeFramePage } from '../TimeFramePage'
+import theme from '../theme'
+import en from '../locales/en.js'
 
 i18n.load('en', { en })
 i18n.activate('en')
 
 const client = {
   readQuery: () => ({
-    impact: JSON.stringify({}),
+    timeFrame: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<ImpactStatementPage />', () => {
+describe('<TimeFramePage />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -28,7 +28,7 @@ describe('<ImpactStatementPage />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <ImpactStatementPage />
+              <TimeFramePage />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
