+++ f2/src/__tests__/WhatHappenedPage.test.js
@@ -1,29 +1,25 @@
 import React from 'react'
+import { i18n } from '@lingui/core'
 import { render, cleanup } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { i18n } from '@lingui/core'
-import { ConfirmationSummary } from '../ConfirmationSummary'
-import theme from '../../theme'
-import en from '../../locales/en.js'
+import { WhatHappenedPage } from '../WhatHappenedPage'
+import theme from '../theme'
+import en from '../locales/en.js'
 
 i18n.load('en', { en })
 i18n.activate('en')
 
 const client = {
   readQuery: () => ({
-    timeFrame: JSON.stringify({}),
     whatHappened: JSON.stringify({}),
-    scammerDetails: JSON.stringify({}),
-    impact: JSON.stringify({}),
-    contactInfo: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-describe('<ConfirmationSummary />', () => {
+describe('<WhatHappenedPage />', () => {
   afterEach(cleanup)
 
   it('renders', () => {
@@ -32,7 +28,7 @@ describe('<ConfirmationSummary />', () => {
         <ThemeProvider theme={theme}>
           <ApolloProvider client={client}>
             <I18nProvider i18n={i18n}>
-              <ConfirmationSummary />
+              <WhatHappenedPage />
             </I18nProvider>
           </ApolloProvider>
         </ThemeProvider>
