+++ f2/src/forms/ImpactStatementInfoForm.js
@@ -6,14 +6,14 @@ import { ApolloConsumer } from 'react-apollo'
 import { useLingui } from '@lingui/react'
 import { Trans } from '@lingui/macro'
 import { Form, Field } from 'react-final-form'
-import { Checkbox } from '../../components/checkbox'
-import { ButtonsContainer } from '../../components/buttons-container'
-import { P } from '../../components/paragraph'
-import { Text } from '../../components/text'
-import { TextArea } from '../../components/text-area'
-import { Label } from '../../components/label'
-import { finalFormAdapter } from '../../utils/finalFormAdapter'
-import { getImpact } from '../../utils/queriesAndMutations'
+import { Checkbox } from '../components/checkbox'
+import { ButtonsContainer } from '../components/buttons-container'
+import { P } from '../components/paragraph'
+import { Text } from '../components/text'
+import { TextArea } from '../components/text-area'
+import { Label } from '../components/label'
+import { finalFormAdapter } from '../utils/finalFormAdapter'
+import { getImpact } from '../utils/queriesAndMutations'
 
 const CheckboxAdapter = finalFormAdapter(Checkbox)
 const TextAreaAdapter = finalFormAdapter(TextArea)
