+++ f2/src/forms/ScammerDetailsForm.js
@@ -4,20 +4,20 @@ import PropTypes from 'prop-types'
 import { css, jsx } from '@emotion/core'
 import { ApolloConsumer } from 'react-apollo'
 import { Form, Field } from 'react-final-form'
-import { Container } from '../../components/container'
+import { Container } from '../components/container'
 import { plural, Trans } from '@lingui/macro'
-import { ButtonsContainer } from '../../components/buttons-container'
-import { TextArea } from '../../components/text-area'
-import { Button } from '../../components/button'
-import { Text } from '../../components/text'
-import { H3 } from '../../components/header'
-import { P } from '../../components/paragraph'
-import { Ul } from '../../components/unordered-list'
-import { Li } from '../../components/list-item'
-import { FileUpload } from '../../components/file-upload'
-import { finalFormAdapter } from '../../utils/finalFormAdapter'
-import { getScammerDetails } from '../../utils/queriesAndMutations'
-import upload from '../../images/upload.svg'
+import { ButtonsContainer } from '../components/buttons-container'
+import { TextArea } from '../components/text-area'
+import { Button } from '../components/button'
+import { Text } from '../components/text'
+import { H3 } from '../components/header'
+import { P } from '../components/paragraph'
+import { Ul } from '../components/unordered-list'
+import { Li } from '../components/list-item'
+import { FileUpload } from '../components/file-upload'
+import { finalFormAdapter } from '../utils/finalFormAdapter'
+import { getScammerDetails } from '../utils/queriesAndMutations'
+import upload from '../images/upload.svg'
 
 const TextAreaAdapter = finalFormAdapter(TextArea)
