+++ f2/src/forms/__tests__/ImpactStatementInfoForm.test.js
@@ -1,44 +1,44 @@
 import React from 'react'
-import { i18n } from '@lingui/core'
-import { MemoryRouter } from 'react-router-dom'
 import wait from 'waait'
+import { i18n } from '@lingui/core'
 import { render, fireEvent, cleanup } from '@testing-library/react'
+import { MemoryRouter } from 'react-router-dom'
 import { MockedProvider } from 'react-apollo/test-utils'
 import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
-import { ContactInfoForm } from '../ContactInfoForm'
-import en from '../../../locales/en.js'
-import theme from '../../../theme'
+import { ImpactStatementInfoForm } from '../ImpactStatementInfoForm'
+import en from '../../locales/en.js'
+import theme from '../../theme'
 
 i18n.load('en', { en })
 i18n.activate('en')
 
 const client = {
   readQuery: () => ({
-    contactInfo: JSON.stringify({}),
+    impact: JSON.stringify({}),
   }),
   writeData: jest.fn(),
 }
 
-const fillIn = (element, { with: value }) =>
-  fireEvent.change(element, { target: { value } })
+// const fillIn = (element, { with: value }) =>
+//   fireEvent.change(element, { target: { value } })
 
 const clickOn = element => fireEvent.click(element)
 
-describe('<ContactInfoForm />', () => {
+describe('<ImpactStatementInfoForm />', () => {
   afterEach(cleanup)
 
-  it.only('calls the onSubmit function when the form is submitted', async () => {
+  it('calls the onSubmit function when the form is submitted', async () => {
     const submitMock = jest.fn()
 
-    const { getAllByRole, getByRole } = render(
+    const { getByRole } = render(
       <MemoryRouter initialEntries={['/']}>
         <ThemeProvider theme={theme}>
           <MockedProvider mocks={[]} addTypename={false}>
             <I18nProvider i18n={i18n}>
               <ApolloProvider client={client}>
-                <ContactInfoForm onSubmit={submitMock} />
+                <ImpactStatementInfoForm onSubmit={submitMock} />
               </ApolloProvider>
             </I18nProvider>
           </MockedProvider>
@@ -46,10 +46,8 @@ describe('<ContactInfoForm />', () => {
       </MemoryRouter>,
     )
 
-    const inputNode = getAllByRole('textbox')[0]
     const nextButton = getByRole('button')
 
-    fillIn(inputNode, { with: 'Mallory' })
     clickOn(nextButton)
     await wait(0) // Wait for promises to resolve
