+++ f2/src/forms/__tests__/TimeFrameInfoForm.test.js
@@ -8,8 +8,8 @@ import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
 import { TimeFrameInfoForm } from '../TimeFrameInfoForm'
-import en from '../../../locales/en.js'
-import theme from '../../../theme'
+import en from '../../locales/en.js'
+import theme from '../../theme'
 
 i18n.load('en', { en })
 i18n.activate('en')
