+++ f2/src/forms/__tests__/WhatHappenedForm.test.js
@@ -8,8 +8,8 @@ import { ApolloProvider } from 'react-apollo'
 import { ThemeProvider } from 'emotion-theming'
 import { I18nProvider } from '@lingui/react'
 import { WhatHappenedForm } from '../WhatHappenedForm'
-import en from '../../../locales/en.js'
-import theme from '../../../theme'
+import en from '../../locales/en.js'
+import theme from '../../theme'
 
 i18n.load('en', { en })
 i18n.activate('en')
